forge-app/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.0.0",
    "arxiv>=2.1.3",
    "ddgs>=9.11.1",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.111.0",
    "openai>=2.24.0",
    "parallel-web>=0.0.1",
    "pydantic>=2.7.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=16.0",
]
